separate `Player` class to a new file
